Lesson 1-What is SRE/exercise2/requirements.txt
flask==1.0.2
pyyaml==3.13
prettytable==0.7.2
